[WIP] impl function `draw_line3`

+++ src/data/i32x2.c
@@ -7,7 +7,7 @@
 
 t_i32x2 encode_i32x2(int32_t x, int32_t y)
 {
-	return ((t_i32x2) ((uint64_t)(uint32_t)x << 32) | (uint32_t)y);
+	return ((t_i32x2) ((uint64_t) x << 32) | (uint32_t)y);
 }
 
 int32_t decode_int_x(t_i32x2 data)
@@ -68,8 +68,6 @@ bool t_i32x2_eq(t_i32x2 a,t_i32x2 b)
 		decode_int_y(a) == decode_int_y(b)
 	);
 }
-
-
 
 /// 定数倍
 t_i32x2 t_i32x2_scalar(int32_t a, t_i32x2 b)
